fixes unit tests for TokenBurnTransitionLogic
