+++ hedera-node/src/test/java/com/hedera/services/txns/token/TokenBurnTransitionLogicTest.java
@@ -60,6 +60,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.verify;
+import static org.mockito.Mockito.doNothing;
 
 @ExtendWith(MockitoExtension.class)
 class TokenBurnTransitionLogicTest {
@@ -70,74 +71,22 @@ class TokenBurnTransitionLogicTest {
 	private final Account treasury = new Account(treasuryId);
 
 	@Mock
-	private Token token;
-	@Mock
-	private TypedTokenStore store;
-	@Mock
 	private TransactionContext txnCtx;
-	@Mock
-	private PlatformTxnAccessor accessor;
+
 	@Mock
 	private OptionValidator validator;
 	@Mock
-	private AccountStore accountStore;
-	@Mock
 	private GlobalDynamicProperties dynamicProperties;
 
-	private TokenRelationship treasuryRel;
 	private TransactionBody tokenBurnTxn;
 
 	private TokenBurnTransitionLogic subject;
+	@Mock
+	private BurnLogic burnLogic;
 
 	@BeforeEach
 	private void setup() {
-		subject = new TokenBurnTransitionLogic(validator, accountStore, store, txnCtx, dynamicProperties);
-	}
-
-	@Test
-	void followsHappyPathForCommon() {
-		// setup:
-		treasuryRel = new TokenRelationship(token, treasury);
-
-		givenValidTxnCtx();
-		given(accessor.getTxn()).willReturn(tokenBurnTxn);
-		given(txnCtx.accessor()).willReturn(accessor);
-		given(store.loadToken(id)).willReturn(token);
-		given(token.getTreasury()).willReturn(treasury);
-		given(store.loadTokenRelationship(token, treasury)).willReturn(treasuryRel);
-		given(token.getType()).willReturn(TokenType.FUNGIBLE_COMMON);
-		// when:
-		subject.doStateTransition();
-
-		// then:
-		verify(token).burn(treasuryRel, amount);
-		verify(store).commitToken(token);
-		verify(store).commitTokenRelationships(List.of(treasuryRel));
-	}
-
-	@Test
-	void followsHappyPathForUnique() {
-		// setup:
-		treasuryRel = new TokenRelationship(token, treasury);
-
-		givenValidUniqueTxnCtx();
-		given(accessor.getTxn()).willReturn(tokenBurnTxn);
-		given(txnCtx.accessor()).willReturn(accessor);
-		given(store.loadToken(id)).willReturn(token);
-		given(token.getTreasury()).willReturn(treasury);
-		given(store.loadTokenRelationship(token, treasury)).willReturn(treasuryRel);
-		given(token.getType()).willReturn(TokenType.NON_FUNGIBLE_UNIQUE);
-		// when:
-		subject.doStateTransition();
-
-		// then:
-		verify(token).getType();
-		verify(store).loadUniqueTokens(token, tokenBurnTxn.getTokenBurn().getSerialNumbersList());
-		verify(token).burn(any(OwnershipTracker.class), eq(treasuryRel), any(List.class));
-		verify(store).commitToken(token);
-		verify(store).commitTokenRelationships(List.of(treasuryRel));
-		verify(store).commitTrackers(any(OwnershipTracker.class));
-		verify(accountStore).commitAccount(any(Account.class));
+		subject = new TokenBurnTransitionLogic(validator, txnCtx, dynamicProperties, burnLogic);
 	}
 
 	@Test
